rpa/auto_esafe/requirements.txt
pyautogui
opencv-python
psutil
python-dotenv
screeninfo
pywin32
requests
pandas
xlrd
openpyxl
